Time: 0 ms (100.00%), Space: 7.9 MB (20.49%) - LeetHub

+++ 2351-first-letter-to-appear-twice/2351-first-letter-to-appear-twice.c
@@ -1,10 +1,10 @@
 char repeatedCharacter(char* s) {
     uint32_t bitmap = 0;
     for (int i = 0; i < strlen(s); i++) {
-        if (bitmap & (1 << (int) (s[i] - 'a'))) {
+        if (bitmap & (1 << (s[i] - 'a'))) {
             return s[i];
         }
-        bitmap = bitmap | (1 << (int) (s[i] - 'a'));
+        bitmap = bitmap | (1 << (s[i] - 'a'));
     }
     return s[0];
 }
